+++ Assets/Plugins/StreamChat/Core/Requests/ReactionRequest.cs
@@ -13,7 +13,7 @@ namespace StreamChat.Core.Requests
         /// <summary>
         /// Reaction score. If not specified reaction has score of 1
         /// </summary>
-        public double? Score { get; set; }
+        public int? Score { get; set; }
 
         /// <summary>
         /// The type of reaction (e.g. 'like', 'laugh', 'wow')
